+++ src/main/java/com/epam/reportportal/cucumber/Utils.java
@@ -42,7 +42,6 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.net.URI;
 import java.util.*;
-import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -123,6 +122,10 @@ public class Utils {
 		rq.setAttributes(attributes);
 		rq.setStartTime(Calendar.getInstance().getTime());
 		rq.setType(type);
+		if ("STEP".equals(type)) {
+			rq.setTestCaseId(TestCaseIdUtils.getTestCaseId(codeRef, null).getId());
+		}
+
 		return rp.startTestItem(rootItemId, rq);
 	}
 
@@ -288,26 +291,6 @@ public class Utils {
 		}
 	}
 
-	@Nullable
-	public static TestCaseIdEntry getTestCaseId(TestStep testStep, String codeRef) {
-		Field definitionMatchField = getDefinitionMatchField(testStep);
-		if (definitionMatchField != null) {
-			try {
-				Method method = retrieveMethod(definitionMatchField, testStep);
-				TestCaseId testCaseIdAnnotation = method.getAnnotation(TestCaseId.class);
-				return ofNullable(testCaseIdAnnotation).flatMap(annotation -> ofNullable(getTestCaseId(testCaseIdAnnotation,
-						method,
-						((PickleStepTestStep) testStep).getDefinitionArgument()
-				)))
-						.orElseGet(() -> getTestCaseId(codeRef, ((PickleStepTestStep) testStep).getDefinitionArgument()));
-			} catch (NoSuchFieldException | IllegalAccessException e) {
-				return getTestCaseId(codeRef, ((PickleStepTestStep) testStep).getDefinitionArgument());
-			}
-		} else {
-			return getTestCaseId(codeRef, ((PickleStepTestStep) testStep).getDefinitionArgument());
-		}
-	}
-
 	static List<ParameterResource> getParameters(List<Argument> arguments, String text) {
 		List<ParameterResource> parameters = Lists.newArrayList();
 		ArrayList<String> parameterNames = Lists.newArrayList();
@@ -345,24 +328,31 @@ public class Utils {
 		return (Method) methodField.get(javaStepDefinition);
 	}
 
-	@Nullable
-	private static TestCaseIdEntry getTestCaseId(TestCaseId testCaseId, Method method, List<Argument> arguments) {
-		if (testCaseId.parametrized()) {
-			return TestCaseIdUtils.getParameterizedTestCaseId(method, arguments.stream().map(Argument::getValue).toArray());
-		} else {
-			return new TestCaseIdEntry(testCaseId.value());
+	private static final java.util.function.Function<List<Argument>, List<?>> ARGUMENTS_TRANSFORM = arguments -> ofNullable(arguments).map(
+			args -> args.stream().map(Argument::getValue).collect(Collectors.toList())).orElse(null);
+
+	@SuppressWarnings("unchecked")
+	public static TestCaseIdEntry getTestCaseId(TestStep testStep, String codeRef) {
+		Field definitionMatchField = getDefinitionMatchField(testStep);
+		List<Argument> arguments = ((PickleStepTestStep) testStep).getDefinitionArgument();
+		if (definitionMatchField != null) {
+			try {
+				Method method = retrieveMethod(definitionMatchField, testStep);
+				return TestCaseIdUtils.getTestCaseId(method.getAnnotation(TestCaseId.class),
+						method,
+						codeRef,
+						(List<Object>) ARGUMENTS_TRANSFORM.apply(arguments)
+				);
+			} catch (NoSuchFieldException | IllegalAccessException ignore) {
+			}
 		}
+		return getTestCaseId(codeRef, arguments);
 	}
 
+	@SuppressWarnings("unchecked")
 	private static TestCaseIdEntry getTestCaseId(String codeRef, List<Argument> arguments) {
-		return ofNullable(arguments).filter(args -> !args.isEmpty())
-				.map(args -> new TestCaseIdEntry(codeRef + TRANSFORM_PARAMETERS.apply(args)))
-				.orElseGet(() -> new TestCaseIdEntry(codeRef));
+		return TestCaseIdUtils.getTestCaseId(codeRef, (List<Object>) ARGUMENTS_TRANSFORM.apply(arguments));
 	}
-
-	private static final Function<List<Argument>, String> TRANSFORM_PARAMETERS = it -> it.stream()
-			.map(Argument::getValue)
-			.collect(Collectors.joining(",", "[", "]"));
 
 	private static Field getDefinitionMatchField(TestStep testStep) {
 		Class<?> clazz = testStep.getClass();