+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- codeRef reporting was added for every item in an item tree
 
 ## [0.0.1-ALPHA]
 ## Added