+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Test Case ID support
 ### Fixed
 - codeRef reporting was added for every item in an item tree