+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.0.2-ALPHA]
 ### Added
 - Test Case ID support
 ### Fixed